+++ tests/test_oauth2_reddit.py
@@ -32,6 +32,20 @@ class OAuth2RedditTest(PRAWTest):
                     'state': '...'}
         self.assertEqual(expected, params)
 
+    # @betamax() is currently broken for this test
+    def test_auto_refresh_token(self):
+        self.r.refresh_access_information(self.refresh_token['identity'])
+        old_token = self.r.access_token
+
+        self.r.config.oauth_auto_refresh = 0.1
+        self.r.user.refresh()
+        current_token = self.r.access_token
+        self.assertNotEqual(old_token, current_token)
+
+        self.r.config.oauth_auto_refresh = 100
+        self.r.user.refresh()
+        self.assertEqual(current_token, self.r.access_token)
+
     @betamax()
     def test_get_access_information(self):
         # If this test fails, the following URL will need to be visted in order